# specs/002-task-api/contracts/task-api.openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: RESTful API for managing user tasks with JWT authentication and user isolation
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@example.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for a user
      description: Retrieve a paginated list of tasks belonging to the authenticated user
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match authenticated user)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum number of tasks to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of tasks to skip (for pagination)
      responses:
        '200':
          description: Successful response with task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  total:
                    type: integer
                    description: Total number of tasks for this user
                  limit:
                    type: integer
                    description: Requested limit
                  offset:
                    type: integer
                    description: Requested offset
                  has_more:
                    type: boolean
                    description: Whether more tasks exist beyond this page
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      operationId: createTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match authenticated user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get task details
      description: Retrieve details of a specific task
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match authenticated user)
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Successful response with task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update a task
      description: Update title and description of an existing task
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match authenticated user)
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete a task
      description: Permanently delete a task
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match authenticated user)
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the is_completed status of a task (true ↔ false)
      operationId: toggleTaskComplete
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match authenticated user)
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (1-500 characters)
          example: "Complete project documentation"
        description:
          type: string
          maxLength: 5000
          default: ""
          description: Task description (optional, 0-5000 characters)
          example: "Write comprehensive API documentation for the task management endpoints"

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated task title (1-500 characters)
          example: "Complete project documentation and review"
        description:
          type: string
          maxLength: 5000
          description: Updated task description (0-5000 characters)
          example: "Write comprehensive API documentation and have it peer-reviewed"

    TaskResponse:
      type: object
      required:
        - id
        - title
        - description
        - is_completed
        - created_at
        - updated_at
        - user_id
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 42
        title:
          type: string
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          description: Task description
          example: "Write comprehensive API documentation for the task management endpoints"
        is_completed:
          type: boolean
          description: Task completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
          example: "2026-02-05T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: "2026-02-05T14:45:00Z"
        user_id:
          type: integer
          description: Owner user ID
          example: 123

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Access denied: cannot access resources of other users"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Location of the validation error (field path)
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type
          example:
            - loc: ["body", "title"]
              msg: "ensure this value has at least 1 characters"
              type: "value_error.any_str.min_length"

  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid authentication credentials"

    Forbidden:
      description: User ID in path does not match authenticated user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Access denied: cannot access resources of other users"

    NotFound:
      description: Task not found or does not belong to authenticated user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

tags:
  - name: Tasks
    description: Task CRUD operations with user isolation